QueryBuilder - split to mutliple files
