+++ src/GraphQL.Common/Request/Builder/QueryBuilder.cs
@@ -1,32 +1,10 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
-using System.Reflection;
 using System.Text;
 using Humanizer;
 
 namespace GraphQL.Common.Request.Builder
 {
-	public interface IQueryBuilder
-	{
-		string Name { get; set; }
-		IQueryBuilder Parent { get; set; }
-		IQueryBuilder CurrentField { get; set; }
-		string Build();
-		IQueryBuilder TryAddField(IQueryBuilder field);
-	}
-
-	public interface IQueryBuilder<out TEntity> : IQueryBuilder
-	{
-
-	}
-
-	public interface IQueryBuilder<out TEntity, out TProp> : IQueryBuilder<TEntity>
-	{
-
-	}
-
 	public abstract class QueryBuilder : IQueryBuilder
 	{
 		protected readonly List<IQueryBuilder> _fields = new List<IQueryBuilder>();
@@ -37,9 +15,8 @@ namespace GraphQL.Common.Request.Builder
 
 		public IQueryBuilder CurrentField { get; set; }
 
-		public QueryBuilder()
+		protected QueryBuilder()
 		{
-
 		}
 
 		protected QueryBuilder(QueryBuilder source)
@@ -83,70 +60,6 @@ namespace GraphQL.Common.Request.Builder
 		}
 	}
 
-	public static class QueryBuilderExtensions
-	{
-		public static IQueryBuilder<TEntity, TProp> Include<TEntity, TProp>(this IQueryBuilder<TEntity> builder, Expression<Func<TEntity, TProp>> propertyExpression)
-		{
-			var body = propertyExpression.Body as MemberExpression;
-			var property = body?.Member as PropertyInfo;
-			if (property == null)
-				throw new ArgumentException("Expression must indicate a property.", nameof(propertyExpression));
-			var propertyName = body.Member.Name;
-
-			var newBuilder = new QueryBuilder<TEntity, TProp>((QueryBuilder<TEntity>) builder);
-			IQueryBuilder field = new QueryBuilder<TEntity, TProp>
-			{
-				Name = propertyName,
-				Parent = newBuilder
-			};
-
-			field = newBuilder.TryAddField(field);
-			newBuilder.CurrentField = field;
-
-			return newBuilder;
-		}
-		public static IQueryBuilder<TEntity, TProp> ThenInclude<TEntity, TProp, TChild>(this IQueryBuilder<TEntity, TProp> builder, Expression<Func<TProp, TChild>> propertyExpression)
-		{
-			var body = propertyExpression.Body as MemberExpression;
-			var property = body?.Member as PropertyInfo;
-			if (property == null)
-				throw new ArgumentException("Expression must indicate a property.", nameof(propertyExpression));
-			var propertyName = body.Member.Name;
-
-			var newBuilder = new QueryBuilder<TEntity, TProp>((QueryBuilder<TEntity>) builder);
-			IQueryBuilder field = new QueryBuilder<TEntity, TProp>
-			{
-				Name = propertyName,
-				Parent = newBuilder
-			};
-
-			field = newBuilder.CurrentField.TryAddField(field);
-			newBuilder.CurrentField = field;
-
-			return newBuilder;
-		}
-		public static IQueryBuilder<TEntity, TChild> ThenInclude<TEntity, TProp, TChild>(this IQueryBuilder<TEntity, IEnumerable<TProp>> builder, Expression<Func<TProp, TChild>> propertyExpression)
-		{
-			var body = propertyExpression.Body as MemberExpression;
-			var property = body?.Member as PropertyInfo;
-			if (property == null)
-				throw new ArgumentException("Expression must indicate a property.", nameof(propertyExpression));
-			var propertyName = body.Member.Name;
-
-			var newBuilder = new QueryBuilder<TEntity, TChild>((QueryBuilder<TEntity>) builder);
-			IQueryBuilder field = new QueryBuilder<TEntity, TChild>
-			{
-				Name = propertyName,
-				Parent = builder
-			};
-
-			field = newBuilder.CurrentField.TryAddField(field);
-			newBuilder.CurrentField = field;
-
-			return newBuilder;
-		}
-	}
-
 	public class QueryBuilder<TEntity> : QueryBuilder, IQueryBuilder<TEntity>
 	{
 		private readonly List<QueryParameter> _parameters = new List<QueryParameter>();
@@ -182,11 +95,6 @@ namespace GraphQL.Common.Request.Builder
 
 			stringBuilder.AppendLine();
 		}
-	}
-
-	public class QueryParameter
-	{
-
 	}
 
 	public class QueryBuilder<TEntity, TProp> : QueryBuilder<TEntity>, IQueryBuilder<TEntity, TProp>
