+++ src/GraphQL.Common/Request/Builder/QueryParameter.cs
@@ -1,0 +1,7 @@
+namespace GraphQL.Common.Request.Builder
+{
+	public class QueryParameter
+	{
+
+	}
+}
